Serhii`s commit
FileManager - new class for saving results of game
in FileManager was modified logic. Added 2, 3 and 4th items.
in Menu created class item, because it can be available and unavailable, it has
 name and number.
Classes Student and Person were implemented by "Serializable" for using in binary files. Other properties of these classes were not modified.
Test results:
questions are not changing.