1. New class Question was created
This class stores question and answer fields which are needed for mission generation. Creating ArrayList with Question data type simplifies the check whether the user's answer is correct since every object consists of a question and a right answer to it.

2. A text file "list_of_questions.dat" with questions and answers for missions was added to a src file.

3. New method initQuestions() was added to the FileManager class.

This method opens "list_of_questions.dat" file, reads information from there and fills the ArrayList with questions and answers.

4. Some magic numbers was substituted with constant values.