Danylo's commit
While testing the program, a bag was detected. To clarify the problem: if one user wins in a mission with a particular teacher, the next user can`t play with him/her, since his/her health are thought to be lower or equal to zero. That issue was solved successfully.

Also, the array of questions is generated only once now: while creating a new player.