Database implementation and conversion to a Apache Maven project.
Instead of using binary files, it was decided to implement the database 'database_java_team' which contains several tables such as: users (to store personal information about users), players (to store information about game progress), teachers (to store information about every boss) and questions_answers (to store bosses tasks).
The database access was made with an object–relational mapping tool Hibernate. Hibernate also requires a set of configuration settings related to database and other related parameters. All such information is supplied with hibernate.cfg.xml file.  The session for connecting to the database is created in HibernateSessionFactoryUtil.class.
Retrieving the data from a database requires classes with annotation @Entity so that Hibernate could create objects with the data from tables on its own. Such classes are Question.class (contains the question, options and right answer), User.class (contains the personal information) and Player.class (contains game progress). Other annotations for example @Table (makes a relation to a particular table), @Id (refers to a unique primary key of each object), @Column (connects a class field to a column name in database) are also used in entity classes.
The functional part of accessing to a database and manipulating with its data are created in the interface DaoImpl.class and implemented by classes for each entity. Each class has methods findById (to get an entity object by its unique id), save (to add new entity object to a table), update (to change the data of existing entity object), delete (to erase an entity object from the table) and findAll (to retrieve all objects of particular entity).
Due to the fact that Teacher.class is abstract and all its successors are a part of one table it is impossible to declarate them as entities. That`s why JDBS was used to retrieve the data with SQLite-requests. All mrthods are implemented in SQLiteJDBC class. It contains such methods as getConnection (to connect to a database), getTeachersName (to get all bosses names) and teacherInfo (to retrieve the data of a particular boss by its name).

The project was converted to a Maven (framework for automating the assembly of projects based on the description of their structure in files in the POM language) project  for the further usages in project visualization, testing etc.