Authorization and registration
Authorization (implemented in AuthorizationController.class): now the program requires the login and password to enter the created account. Method getUserByLogin() checks whether the login is valid and stored in database. If so, the method checkByPassword() compares the entered password with one that stored in a database. On condition that they are equal, the user can enter its account.

Registration(implemented in RegistrationController.class: The registration process was modified by implementation of classes Matcher and Pattern, whose methods check the input according to the regexes. All regexes are stored in property file regex.properties. Also, the program doesn't allow to enter logins and emails that are already stored in a database.